Render token as code and simplify a bit
Show the token as monospaced code rather than normal text, because I find it easier to select and copy without having an extra space on the end. Remove the quotation marks from the token (there's that regex again).

Make the executor synchronous (since its body is entirely synchronous, no need for async) and remove the `args` argument since it is not used at all.

Switch "we" to "I" in the message when the token is not found, so that it gives the user the impression of the singular client acting on its own, rather than some organization getting the token and watching.

Also make the use of quotes consistent, using all double quotes and getting rid of single quotes and backticks where applicable. (OCD again.)